WIP WIP WIP WIP
I've changed the urls for the view to be /wishlist/{steam_id}.
I've added consts to use as URLS and build the requests.
I've started writing the helper functions to use in the retrieving of the wishlist and deals.
Also I've added a concept of "deal-source" which will be an expandable sources of deals to look for products in.